+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Upgrade to v2.3.0 of the Cloudflare Terraform Provider
+* Upgrade to pulumi-terraform-bridge v1.5.2
 
 ---